@@ -3,33 +3,33 @@ version: 0.0.2
 description: "A clean & simple Catan player & game data manager"
 
 dependencies:
-  bloc: ^6.1.3
-  dartz: ^0.9.2
+  bloc: ^7.0.0-nullsafety.4
+  dartz: ^0.10.0-nullsafety.1
   datetime_picker_formfield: ^2.0.0
   equatable: ^2.0.0
   flutter:
     sdk: flutter
-  flutter_bloc: ^6.1.3
-  fluttertoast: ^7.1.8
-  freezed_annotation: ^0.14.0
-  hive: ^1.4.4
-  hive_flutter: ^0.3.1
+  flutter_bloc: ^7.0.0-nullsafety.5
+  fluttertoast: ^8.0.1-nullsafety.0
+  freezed_annotation: ^0.14.1
+  hive: ^1.6.0-nullsafety.2
+  hive_flutter: ^1.0.0-nullsafety.0
   intl: ^0.17.0
   json_annotation: ^4.0.0
-  polygon_clipper: ^1.0.2
-  provider: ^4.3.3
+  provider: ^5.0.0
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
   build_runner:
-  flutter_launcher_icons: ^0.8.0
+  flutter_launcher_icons: ^0.9.0
   flutter_test:
     sdk: flutter
   hive_generator: ^0.8.2
   json_serializable: ^4.0.2
-  test: ^1.15.7
+  test: ^1.16.5
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 flutter_icons:
   android: launcher_icon
