@@ -7,16 +7,16 @@ class LoadPlayers extends PlayerEvent {}
 
 class AddOrUpdatePlayer extends PlayerEvent {
 
-  final Player toEdit;
+  final Player? toEdit;
   final String name;
   final Gender gender;
   final Color color;
 
-  AddOrUpdatePlayer({this.toEdit, @required this.name, @required this.gender, @required this.color});
+  AddOrUpdatePlayer({this.toEdit, required this.name, required this.gender, required this.color});
 
-  AddOrUpdatePlayer.add({@required this.name, @required this.gender, @required this.color}) : toEdit = null;
+  AddOrUpdatePlayer.add({required this.name, required this.gender, required this.color}) : toEdit = null;
 
-  AddOrUpdatePlayer.edit({@required this.toEdit, @required this.name, @required this.gender, @required this.color});
+  AddOrUpdatePlayer.edit({required this.toEdit, required this.name, required this.gender, required this.color});
 
   bool get isEdit => toEdit != null;
 }
@@ -25,5 +25,5 @@ class DeletePlayerEvent extends PlayerEvent {
 
   final Player player;
 
-  DeletePlayerEvent(this.player) : assert(player != null);
+  DeletePlayerEvent(this.player);
 }
