@@ -1,16 +1,16 @@
 import 'dart:ui';
 
+import 'package:catan_master/presentation/core/polygon/polygon_border.dart';
 import 'package:flutter/material.dart';
-import 'package:polygon_clipper/polygon_border.dart';
 
 /// TODO fix width / height / rotation to tightly match
 class Hexagon extends StatelessWidget {
 
-  final Color color;
+  final Color? color;
   final double width;
   final double height;
   final double rotate;
-  final Widget child;
+  final Widget? child;
   final List<BoxShadow> shadows;
 
   Hexagon({
