@@ -8,12 +8,12 @@ class UserFeedback extends StatelessWidget {
 
   final Widget child;
 
-  UserFeedback({@required this.child});
+  UserFeedback({required this.child});
 
   @override
   Widget build(BuildContext context) {
     return BlocListener<FeedbackBloc, FeedbackState>(
-      listenWhen: (_s1, _s2) => ModalRoute.of(context).isCurrent,
+      listenWhen: (_s1, _s2) => ModalRoute.of(context)?.isCurrent ?? false,
       listener: (context, FeedbackState state) {
         FeedbackMessage feedback = state.last;
         feedback.show(context);
