@@ -5,10 +5,10 @@ import 'package:flutter/material.dart';
 class GameActions extends StatelessWidget {
 
   final Game game;
-  final Function onEdit;
-  final Function onDelete;
+  final VoidCallback onEdit;
+  final VoidCallback onDelete;
 
-  GameActions({@required this.game, @required this.onEdit, @required this.onDelete});
+  GameActions({required this.game, required this.onEdit, required this.onDelete});
 
   @override
   Widget build(BuildContext context) {
