@@ -24,17 +24,20 @@ class PlayerStatsPage extends StatelessWidget {
       var statistics = state.getStatisticsForPlayer(player);
       return NestedScrollView(
         headerSliverBuilder: (context, innerBoxIsScrolled) {
+          Brightness brightness = ThemeData.estimateBrightnessForColor(player.color);
+          var theme = ThemeData(brightness: brightness);
           return <Widget>[
             SliverOverlapAbsorber(
               handle: NestedScrollView.sliverOverlapAbsorberHandleFor(context),
               sliver: SliverAppBar(
                 centerTitle: true,
-                flexibleSpace: FlexibleSpaceBar(title: Text(player.name), centerTitle: true,),
+                flexibleSpace: FlexibleSpaceBar(title: Text(player.name, style: theme.textTheme.headline6), centerTitle: true,),
                 collapsedHeight: kToolbarHeight,
                 toolbarHeight: kToolbarHeight,
                 expandedHeight: 100.0,
                 pinned: true,
                 backgroundColor: player.color,
+                iconTheme: brightness == Brightness.light ? IconTheme.of(context).copyWith(color: Colors.black) : null,
                 actions: [
                   IconButton(
                     onPressed: () => Navigator.of(context).pushNamed("/players/edit", arguments: {"player": player}),
@@ -98,17 +101,17 @@ class PlayerStatsPage extends StatelessWidget {
               HorizontalInfoTile(
                 leading: Icon(Icons.history),
                 start: Text("Last Game"),
-                end: Text(statistics.lastGame == null ? "TBD" : DateFormat.yMd().format(statistics.lastGame.date)),
+                end: Text(statistics.lastGame == null ? "TBD" : DateFormat.yMd().format(statistics.lastGame!.date)),
               ),
               HorizontalInfoTile(
                 leading: statistics.mostPlayedExpansion.iconWidget,
                 start: Text("Most Played"),
-                end: Text(statistics.mostPlayedExpansion.name ),
+                end: Text(statistics.mostPlayedExpansion.name! ),
               ),
               HorizontalInfoTile(
                 leading: statistics.mostWonExpansion.iconWidget,
                 start: Text("Most Won"),
-                end: Text(statistics.mostWonExpansion.name),
+                end: Text(statistics.mostWonExpansion.name!),
               ),
               HorizontalInfoTile(
                 leading: Icon(Icons.favorite_outline),
@@ -187,22 +190,23 @@ class AchievementLine extends StatelessWidget {
   }
 }
 
-extension AchievementUi on Prize {
+extension PrizeUi on Prize {
 
-  IconData get icon {
+  IconData? get icon {
       switch (this.type) {
         case PrizeType.expansion_master:
           return (value as CatanExpansion).icon;
         case PrizeType.on_a_roll:
           return CatanIcons.dice;
+        //TODO add more icons
         case PrizeType.catan_addict:
         case PrizeType.newbie:
         case PrizeType.loser:
       }
-      return Icons.help_outline;
+      return null;
   }
 
-  Widget get leadingWidget => this == null ? Hexagon() : Icon(icon);
+  Widget get leadingWidget => icon == null ? Hexagon() : Icon(icon);
 
   String get title {
     switch (this.type) {
@@ -234,8 +238,8 @@ extension AchievementUi on Prize {
 
 class DetailLine extends StatelessWidget {
 
-  final IconData icon;
-  final Widget child;
+  final IconData? icon;
+  final Widget? child;
 
   DetailLine({this.icon, this.child});
 
@@ -245,7 +249,7 @@ class DetailLine extends StatelessWidget {
       children: [
         Icon(icon),
         SizedBox(width: 16.0,),
-        child,
+        child!,
       ],
     );
   }
