@@ -23,7 +23,7 @@ class GameAdded extends GamesLoaded {
 
   final Game newGame;
 
-  GameAdded(Games games, {this.newGame}) : super(Games(List.from(games.games)..add(newGame)));
+  GameAdded(Games games, {required this.newGame}) : super(Games(List.from(games.games)..add(newGame)));
 
 }
 
@@ -32,6 +32,6 @@ class GameEdited extends GamesLoaded {
   //TODO move to games
   final Game editedGame;
 
-  GameEdited(Games games, {this.editedGame}) : super(Games(List.from(games.games)..add(editedGame)));
+  GameEdited(Games games, {required this.editedGame}) : super(Games(List.from(games.games)..add(editedGame)));
 
 }
