@@ -16,8 +16,8 @@ extension ShowFeedback on FeedbackMessage {
           action: hasAction
               ? SnackBarAction(
                   textColor: severity == Severity.error ? Colors.red : null,
-                  label: action.text,
-                  onPressed: action.action
+                  label: action!.text,
+                  onPressed: action!.action
               )
               : null,
         ));
@@ -26,7 +26,7 @@ extension ShowFeedback on FeedbackMessage {
         showDialog(
             context: context,
             builder: (context) => AlertDialog(
-              title: Text(title),
+              title: Text(title!),
               content: Text(message),
               actions: [
                 if (!hasAction) TextButton(onPressed: () => Navigator.of(context).pop(), child: Text("OK")),
